+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0a4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Add converter for fileupload to namedblobfile.
+  [jensens]
 
 
 1.0a3 (2021-07-09)